+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Enhancements
   * make regexp validator support list
   * support mount alias modules
+  * add `all_or_none_of` validation
 * Bugfix
   * Support :chunk response